requirements.txt
gevent==21.1.2
greenlet==1.0.0
web3==5.15.0
pysqlcipher3==1.0.3
requests==2.25.1
pycryptodome==3.6.6
pycryptodomex==3.6.6
coincurve==13.0.0
typing-extensions==3.7.4.3
base58check==1.0.2
bech32==1.2.0
gql==2.0.0
substrate-interface==0.11.16
beautifulsoup4==4.9.3

# For the rest api
flask-cors==3.0.8
flask-restful==0.3.8
flask==1.1.1
marshmallow==3.2.1
webargs==6.0.0

#constraints
setuptools==41.0.1  # constraint required by pyinstaller==3.5.0

# substrate-indexer
gevent-websocket==0.10.1  # TODO: move up, as it is general
flask-socketio==5.0.1
